Bugfix: Volume should be stored as a float.
Volume was being stored as an integer in the database, so whenever a float was passed in it was rounding it down to zero.

That may be why our javascript and our crud app are not synching up and why several attributes that should be floats are showing up as zero.  They keep getting rounded down to 0 because they are floats.  When volume is rounded down to zero, no sound appears to come out of the speakers and our app breaks.

+++ db/migrate/20140314155358_create_keys.rb
@@ -7,7 +7,7 @@ class CreateKeys < ActiveRecord::Migration
       t.integer :panner
       t.integer :delay
       t.integer :note_filter
-      t.integer :volume
+      t.float   :volume
       t.integer :key_code
 
       t.timestamps
